+++ facebook/__init__.py
@@ -45,7 +45,7 @@ __version__ = version.__version__
 
 FACEBOOK_GRAPH_URL = "https://graph.facebook.com/"
 FACEBOOK_OAUTH_DIALOG_URL = "https://www.facebook.com/dialog/oauth?"
-VALID_API_VERSIONS = ["2.4", "2.5", "2.6", "2.7", "2.8", "2.9"]
+VALID_API_VERSIONS = ["2.4", "2.5", "2.6", "2.7", "2.8", "2.9", "2.10"]
 VALID_SEARCH_TYPES = ["page", "event", "group", "place", "placetopic", "user"]
 
 
@@ -89,7 +89,7 @@ class GraphAPI(object):
         self.session = session or requests.Session()
 
         if version:
-            version_regex = re.compile("^\d\.\d$")
+            version_regex = re.compile("^\d\.\d{1,2}$")
             match = version_regex.search(str(version))
             if match is not None:
                 if str(version) not in VALID_API_VERSIONS:
@@ -229,7 +229,7 @@ class GraphAPI(object):
         try:
             headers = response.headers
             version = headers["facebook-api-version"].replace("v", "")
-            return float(version)
+            return str(version)
         except Exception:
             raise GraphAPIError("API version number not available")