+++ docs/changes.rst
@@ -5,7 +5,7 @@ Changelog
 Version 3.0.0 (unreleased)
 ========================
  - Remove support for Python 2.6.
- - Add support for Graph API version 2.8.
+ - Add support for Graph API versions  2.8 and 2.9.
  - Remove support for Graph API versions 2.1 and 2.2.
  - Change default Graph API version to 2.3.
  - Add support for requests' sessions (#201).